+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 2.5.1 (2021-09-15)
+
 ### Added
 
 -   `GetFormSubmissionApproval` to `ApprovalsClient`